Adding shipping charge as floating point

+++ db/migrate/20160608153203_create_shipping_methods.rb
@@ -1,7 +1,7 @@
 class CreateShippingMethods < ActiveRecord::Migration[5.0]
   def change
     create_table :shipping_methods do |t|
-      t.integer :shipping_charge, null: false, default: 0
+      t.float :shipping_charge, null: false, default: 0
       t.integer :delivery_period, null: false, default: 0
       t.references :vendor, index: true
